Allowed multiline CriticMarkup
The CriticMarkup documentation
[states](https://github.com/CriticMarkup/CriticMarkup-toolkit) that
linebreaks should be avoided, but seems to permit them; also the parser
seems to process them well.  (In my opinion, the part on inline
elements probably refers to avoiding paragraph breaks, or something
causing the insertion of line breaking HTML elements.)

For people who are willing to take the risk, it would be nice to
highlight multiline CriticMarkup, which is what these changes do.